Bump Sphinx to 6, myst-parser 3
New attempt at fixing the documentation generation. Bumping the versions
appears to take care of an error about something "applehelp" with
sphinx.  That did show up in a local run, then it didn't show up in an
extra CI run and then it did show up again in the nightly run.

I now at least have a hypothesis: the `database.json` file contains the
commit hashes and Sphinx doesn't run if the commit hashes match.  A
broken build that still updates that file, causes the next run to do
nothing and so appear succesful.

This commits bumps the Sphinx version, which I think will fix the error.
It only bumps to version 6 because version 7 gives an error:

  Local asset file paths must not contain query strings ...

That's a problem for another day.

The database is still corrupt, regenerated documentation is added in the
next commit.

Signed-off-by: Erik Boasson <eb@ilities.com>